requirements.txt
streamlit==1.41.0
openai==1.61.0
chromadb==0.6.3
sentence-transformers==3.4.1
duckduckgo-search==7.3.2
newspaper3k==0.2.8
lxml[html_clean]==5.3.0
python-dotenv==1.0.1
flask==3.1.0
flask-cors==5.0.0
